Rename http_listener to influxdb_listener
