+++ plugins/inputs/http_listener_v2/README.md
@@ -1,21 +1,11 @@
-# Generic HTTP listener service input plugin
+# HTTP Listener v2 Input Plugin
 
-> NOTE: This is a new version of HTTP listener plugin.
-> This plugin supports all [data formats](/docs/DATA_FORMATS_INPUT.md) while the old [http_listener](/plugins/inputs/http_listener)
-> only accepts data in InfluxDB line-protocol only
+HTTP Listener v2 is a service input plugin that listens for metrics sent via
+HTTP.  Metrics may be sent in any supported [data format][data_format].
 
-The HTTP listener is a service input plugin that listens for messages sent via HTTP POST.
-
-Enable TLS by specifying the file names of a service TLS certificate and key.
-
-Enable mutually authenticated TLS and authorize client connections by signing certificate authority by including a list of allowed CA certificate file names in ````tls_allowed_cacerts````.
-
-Enable basic HTTP authentication of clients by specifying a username and password to check for. These credentials will be received from the client _as plain text_ if TLS is not configured.
-
-**Example:**
-```
-curl -i -XPOST 'http://localhost:8080/write' --data-binary 'cpu_load_short,host=server01,region=us-west value=0.64 1434055562000000000'
-```
+**Note:** The plugin previously known as `http_listener` has been renamed
+`influxdb_listener`.  If you would like Telegraf to act as a proxy/relay for
+InfluxDB it is recommended to use [`influxdb_listener`][influxdb_listener].
 
 ### Configuration:
 
@@ -41,7 +31,7 @@ This is a sample configuration for the plugin.
   ## 0 means to use the default of 536,870,912 bytes (500 mebibytes)
   max_body_size = 0
 
-  ## Set one or more allowed client CA certificate file names to 
+  ## Set one or more allowed client CA certificate file names to
   ## enable mutually authenticated TLS connections
   tls_allowed_cacerts = ["/etc/telegraf/clientca.pem"]
 
@@ -60,3 +50,22 @@ This is a sample configuration for the plugin.
   ## https://github.com/influxdata/telegraf/blob/master/docs/DATA_FORMATS_INPUT.md
   data_format = "influx"
 ```
+
+### Metrics:
+
+Metrics are created from the request body and are dependant on the value of `data_format`.
+
+### Troubleshooting:
+
+**Send Line Protocol**
+```
+curl -i -XPOST 'http://localhost:8080/telegraf' --data-binary 'cpu_load_short,host=server01,region=us-west value=0.64 1434055562000000000'
+```
+
+**Send JSON**
+```
+curl -i -XPOST 'http://localhost:8080/telegraf' --data-binary '{"value1": 42, "value2": 42}'
+```
+
+[data_format]: /docs/DATA_FORMATS_INPUT.md
+[influxdb_listener]: /plugins/inputs/influxdb_listener/README.md
